requirements.txt
cassandra-driver==3.25.0
click==8.0.3
futures==2.2.0
geomet==0.2.1.post1
six==1.16.0
